move icons to single file and use new design
