+++ apps/web/vibes/soul/components/footer/index.tsx
@@ -57,11 +57,10 @@ export const Footer = function Footer({
     >
       <div className="mx-auto max-w-screen-2xl">
         <div className="mx-3 flex flex-col justify-between gap-10 border-t border-t-contrast-100 pt-16 @xl:mx-6 @xl:py-20 @2xl:flex-row @5xl:mx-20">
-          {/* Contact Info || Logo */}
-          <div className="@2xl:w-1/3">
+          <div className="flex flex-col @2xl:w-1/3">
             {/* Contact Information */}
             {contactInformation?.address || contactInformation?.phone ? (
-              <div className="text-2xl font-medium">
+              <div className="text-[20px] font-medium @lg:text-2xl">
                 <h3 className="text-contrast-300">Contact Us</h3>
                 <div>
                   {contactInformation?.address && <p>{contactInformation.address}</p>}
@@ -91,13 +90,13 @@ export const Footer = function Footer({
 
             {/* Social Media Links */}
             {socialMediaLinks && (
-              <div className="mt-4 flex items-center">
+              <div className="mt-auto flex items-center gap-4 pt-8">
                 {socialMediaLinks.map(({ href, icon }, i) => {
                   return (
                     <Link
                       key={i}
                       href={href}
-                      className="block rounded-full fill-contrast-300 p-1.5 ring-primary transition-colors duration-300 ease-out hover:fill-foreground focus-visible:outline-0 focus-visible:ring-2"
+                      className="flex items-center justify-center rounded-lg fill-contrast-400 p-1 ring-primary transition-colors duration-300 ease-out hover:fill-foreground focus-visible:outline-0 focus-visible:ring-2"
                     >
                       {icon}
                     </Link>
