+++ apps/web/vibes/soul/components/footer/payment-icons.tsx
@@ -1,0 +1,147 @@
+export const Amex = function Amex() {
+  return (
+    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="#1F72CD" stroke="#D9D9D9" />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M6.68111 8.5L3.5 15.7467H7.30824L7.78035 14.5913H8.85949L9.3316 15.7467H13.5234V14.8649L13.8969 15.7467H16.0652L16.4387 14.8462V15.7467H25.1566L26.2166 14.6213L27.2092 15.7467L31.6868 15.7561L28.4957 12.1436L31.6868 8.5H27.2786L26.2467 9.60463L25.2854 8.5H15.8016L14.9872 10.3704L14.1537 8.5H10.3534V9.35186L9.93068 8.5H6.68111ZM19.6994 9.52905H24.7055L26.2367 11.2316L27.8172 9.52905H29.3484L27.0219 12.1426L29.3484 14.726H27.7477L26.2166 13.0036L24.628 14.726H19.6994V9.52905ZM20.9356 11.555V10.6057V10.6048H24.0593L25.4223 12.1229L23.9989 13.6493H20.9356V12.613H23.6667V11.555H20.9356ZM7.41799 9.52905H9.2743L11.3843 14.4431V9.52905H13.4178L15.0476 13.0524L16.5496 9.52905H18.573V14.7291H17.3418L17.3317 10.6544L15.5368 14.7291H14.4355L12.6305 10.6544V14.7291H10.0977L9.61756 13.5633H7.02338L6.54421 14.728H5.18717L7.41799 9.52905ZM7.46634 12.4856L8.32101 10.4089L9.1747 12.4856H7.46634Z"
+        fill="white"
+      />
+    </svg>
+  )
+}
+
+export const ApplePay = function ApplePay() {
+  return (
+    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="white" stroke="#D9D9D9" />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M8.3462 8.93296C8.81624 8.97261 9.28628 8.69506 9.58006 8.34316C9.86894 7.98135 10.0599 7.49563 10.0109 7C9.59475 7.01983 9.08064 7.27755 8.78686 7.63936C8.51757 7.95161 8.28745 8.45715 8.3462 8.93296ZM13.9377 15.3167V7.58981H16.8021C18.2807 7.58981 19.3138 8.62072 19.3138 10.1274C19.3138 11.6342 18.2611 12.675 16.7629 12.675H15.1226V15.3167H13.9377ZM10.006 9.02714C9.5919 9.00301 9.21406 9.1532 8.90886 9.27451C8.71246 9.35257 8.54614 9.41868 8.41961 9.41868C8.27762 9.41868 8.10444 9.34904 7.90999 9.27085C7.65521 9.16839 7.36392 9.05125 7.05845 9.05687C6.35829 9.06679 5.70709 9.46825 5.34966 10.1076C4.61522 11.3863 5.1587 13.2796 5.86866 14.3205C6.2163 14.8359 6.63248 15.4009 7.18086 15.3811C7.42211 15.3719 7.59566 15.2974 7.77526 15.2202C7.98203 15.1314 8.19683 15.0391 8.53223 15.0391C8.85599 15.0391 9.0614 15.129 9.25858 15.2152C9.44606 15.2973 9.62611 15.376 9.89339 15.3712C10.4614 15.3613 10.8188 14.8557 11.1664 14.3403C11.5416 13.7871 11.7064 13.2471 11.7314 13.1652L11.7344 13.1557C11.7338 13.1551 11.7291 13.153 11.721 13.1492C11.5956 13.0911 10.637 12.6469 10.6278 11.4557C10.6186 10.4559 11.3881 9.94937 11.5093 9.86963L11.5093 9.86962C11.5166 9.86477 11.5216 9.8615 11.5238 9.85979C11.0342 9.12626 10.2704 9.04696 10.006 9.02714ZM21.478 15.3761C22.2223 15.3761 22.9126 14.9945 23.226 14.3898H23.2505V15.3166H24.3472V11.4706C24.3472 10.3554 23.4659 9.63672 22.1096 9.63672C20.8513 9.63672 19.921 10.3653 19.8867 11.3665H20.9541C21.0423 10.8907 21.478 10.5784 22.0754 10.5784C22.8 10.5784 23.2064 10.9204 23.2064 11.5499V11.9761L21.7277 12.0653C20.3519 12.1496 19.6077 12.7195 19.6077 13.7108C19.6077 14.712 20.3764 15.3761 21.478 15.3761ZM21.7962 14.4592C21.1646 14.4592 20.7631 14.1519 20.7631 13.6811C20.7631 13.1954 21.1499 12.9129 21.8893 12.8683L23.2063 12.784V13.2202C23.2063 13.9438 22.5992 14.4592 21.7962 14.4592ZM27.9851 15.619C27.5102 16.9721 26.9667 17.4181 25.8112 17.4181C25.7231 17.4181 25.4293 17.4082 25.3607 17.3884V16.4616C25.4342 16.4715 25.6153 16.4814 25.7084 16.4814C26.2323 16.4814 26.526 16.2584 26.7072 15.6785L26.8149 15.3365L24.8075 9.71108H26.0462L27.4416 14.2758H27.4661L28.8616 9.71108H30.066L27.9851 15.619ZM15.1225 8.60087H16.4886C17.5168 8.60087 18.1044 9.15598 18.1044 10.1324C18.1044 11.1088 17.5168 11.6688 16.4837 11.6688H15.1225V8.60087Z"
+        fill="black"
+      />
+    </svg>
+  )
+}
+
+export const Bitcoin = function Bitcoin() {
+  return (
+    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="white" stroke="#D9D9D9" />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M25.6269 12.0646C25.6269 16.5185 22.0165 20.1292 17.5627 20.1292C13.109 20.1292 9.49854 16.5185 9.49854 12.0646C9.49854 7.61064 13.109 4 17.5627 4C22.0165 4 25.6269 7.61064 25.6269 12.0646ZM19.5001 9.06828C20.6219 9.45351 21.4424 10.0303 21.2815 11.1038C21.1645 11.8899 20.7272 12.2701 20.146 12.4032C20.9433 12.8167 21.215 13.6014 20.9626 14.5502C20.4832 15.9157 19.3437 16.0306 17.8281 15.7452L17.4601 17.213L16.5717 16.9922L16.9346 15.5441C16.7043 15.4871 16.4689 15.4261 16.2265 15.361L15.862 16.8162L14.9746 16.5954L15.3421 15.1245L13.5527 14.6748L13.9941 13.6613C13.9941 13.6613 14.6491 13.8342 14.64 13.8216C14.8916 13.8836 15.0035 13.7203 15.0475 13.6119L16.0447 9.62797C16.0559 9.43989 15.9906 9.20291 15.6307 9.11316C15.6443 9.10358 14.9852 8.95332 14.9852 8.95332L15.2216 8.00789L17.0136 8.4501L17.3781 6.99591L18.266 7.21676L17.9091 8.64221C18.148 8.69616 18.388 8.75112 18.6213 8.80911L18.9757 7.39274L19.8641 7.61359L19.5001 9.06828ZM17.3744 11.4941C17.9796 11.6545 19.2964 12.0037 19.5259 11.0882C19.7601 10.1516 18.4807 9.86878 17.8543 9.73028L17.8543 9.73027C17.7833 9.71457 17.7207 9.70074 17.6697 9.68801L17.2272 11.4558C17.2694 11.4662 17.3189 11.4794 17.3744 11.4941ZM16.6887 14.3404C17.4136 14.5313 18.9988 14.9489 19.251 13.9402C19.5092 12.9088 17.9722 12.5652 17.2223 12.3976C17.1385 12.3789 17.0645 12.3624 17.004 12.3473L16.516 14.2956C16.5656 14.3079 16.6236 14.3232 16.6886 14.3403L16.6887 14.3404Z"
+        fill="#F7931A"
+      />
+    </svg>
+  )
+}
+
+export const GooglePay = function GooglePay() {
+  return (
+    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="white" stroke="#D9D9D9" />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M16.5301 15.7585V12.7375H18.0893C18.7282 12.7375 19.2675 12.5234 19.7071 12.1012L19.8126 11.9942C20.6156 11.12 20.5629 9.75817 19.7071 8.94941C19.2792 8.52125 18.6931 8.28932 18.0893 8.30122H15.5864V15.7585H16.5301ZM16.5303 11.8217V9.21703H18.1131C18.4531 9.21703 18.7756 9.34786 19.0159 9.58573C19.526 10.0853 19.5377 10.9178 19.0453 11.4352C18.8049 11.6909 18.4649 11.8336 18.1131 11.8217H16.5303ZM24.2147 11.0546C23.8102 10.6799 23.2592 10.4896 22.5617 10.4896C21.6649 10.4896 20.9908 10.8227 20.5453 11.4827L21.3777 12.012C21.6825 11.5601 22.0987 11.3341 22.6262 11.3341C22.9603 11.3341 23.2827 11.459 23.5347 11.6849C23.7809 11.899 23.9216 12.2083 23.9216 12.5353V12.7554C23.5582 12.5532 23.101 12.4461 22.5383 12.4461C21.8818 12.4461 21.3542 12.6007 20.9615 12.9159C20.5688 13.2311 20.3695 13.6474 20.3695 14.1766C20.3578 14.6583 20.5629 15.1162 20.9263 15.4254C21.2956 15.7585 21.7645 15.925 22.3155 15.925C22.9662 15.925 23.482 15.6336 23.8747 15.0508H23.9157V15.7585H24.8184V12.6126C24.8184 11.9525 24.6191 11.4292 24.2147 11.0546ZM21.6533 14.8189C21.4599 14.6761 21.3426 14.4442 21.3426 14.1944C21.3426 13.9149 21.4716 13.683 21.7237 13.4987C21.9816 13.3143 22.3041 13.2192 22.6851 13.2192C23.2128 13.2132 23.6231 13.3322 23.9163 13.57C23.9163 13.9744 23.758 14.3253 23.4473 14.6226C23.1659 14.9081 22.7848 15.0686 22.3862 15.0686C22.1223 15.0746 21.8644 14.9854 21.6533 14.8189ZM26.8465 18.0004L30 10.6562H28.9743L27.5147 14.3194H27.4972L26.0025 10.6562H24.9767L27.0458 15.4314L25.8735 18.0004H26.8465Z"
+        fill="#3C4043"
+      />
+      <path
+        d="M13.272 12.0831C13.272 11.7917 13.2486 11.5003 13.2017 11.2148H9.22168V12.8621H11.5018C11.408 13.3914 11.1032 13.8671 10.6578 14.1644V15.2349H12.0176C12.8148 14.4915 13.272 13.3914 13.272 12.0831Z"
+        fill="#4285F4"
+      />
+      <path
+        d="M9.22211 16.2696C10.3592 16.2696 11.3205 15.889 12.0181 15.2349L10.6582 14.1645C10.2772 14.4261 9.79068 14.5748 9.22211 14.5748C8.12013 14.5748 7.18815 13.8195 6.85404 12.8086H5.45312V13.9147C6.16824 15.3598 7.62776 16.2696 9.22211 16.2696Z"
+        fill="#34A853"
+      />
+      <path
+        d="M6.85401 12.8088C6.67814 12.2796 6.67814 11.7027 6.85401 11.1675V10.0674H5.45288C4.84904 11.2746 4.84904 12.7018 5.45288 13.909L6.85401 12.8088Z"
+        fill="#FBBC04"
+      />
+      <path
+        d="M9.22211 9.40114C9.82585 9.38925 10.4061 9.62117 10.8399 10.0434L12.0474 8.81836C11.2795 8.09285 10.2713 7.69442 9.22211 7.70631C7.62776 7.70631 6.16824 8.62211 5.45312 10.0672L6.85404 11.1733C7.18815 10.1564 8.12013 9.40114 9.22211 9.40114Z"
+        fill="#EA4335"
+      />
+    </svg>
+  )
+}
+
+export const Mastercard = function Mastercard() {
+  return (
+    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="white" stroke="#D9D9D9" />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M17.75 17.1569C16.5584 18.1852 15.0128 18.806 13.3238 18.806C9.55511 18.806 6.5 15.7154 6.5 11.903C6.5 8.09057 9.55511 5 13.3238 5C15.0128 5 16.5584 5.62075 17.75 6.64903C18.9416 5.62075 20.4872 5 22.1762 5C25.9449 5 29 8.09057 29 11.903C29 15.7154 25.9449 18.806 22.1762 18.806C20.4872 18.806 18.9416 18.1852 17.75 17.1569Z"
+        fill="#ED0006"
+      />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M17.75 17.1569C19.2172 15.8908 20.1475 14.0068 20.1475 11.903C20.1475 9.79917 19.2172 7.91517 17.75 6.64903C18.9416 5.62075 20.4872 5 22.1762 5C25.9449 5 29 8.09057 29 11.903C29 15.7154 25.9449 18.806 22.1762 18.806C20.4872 18.806 18.9416 18.1852 17.75 17.1569Z"
+        fill="#F9A000"
+      />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M17.75 6.64941C19.2172 7.91555 20.1476 9.79954 20.1476 11.9033C20.1476 14.0071 19.2172 15.8911 17.75 17.1573C16.2829 15.8911 15.3525 14.0071 15.3525 11.9033C15.3525 9.79954 16.2829 7.91555 17.75 6.64941Z"
+        fill="#FF5E00"
+      />
+    </svg>
+  )
+}
+
+export const Paypal = function Paypal() {
+  return (
+    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="white" stroke="#D9D9D9" />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M9.90001 15.2412H8.34439C8.23794 15.2412 8.1474 15.3186 8.1308 15.4236L7.50163 19.4127C7.48913 19.4914 7.55009 19.5624 7.62993 19.5624H8.37259C8.47905 19.5624 8.56958 19.485 8.58618 19.3797L8.75588 18.3038C8.77225 18.1985 8.863 18.1212 8.96924 18.1212H9.46169C10.4864 18.1212 11.0778 17.6253 11.2323 16.6427C11.3019 16.2128 11.2353 15.875 11.0339 15.6384C10.8129 15.3787 10.4206 15.2412 9.90001 15.2412ZM10.0795 16.6982C9.99443 17.2564 9.56792 17.2564 9.15553 17.2564H8.92078L9.08547 16.2138C9.09525 16.1509 9.14984 16.1045 9.21353 16.1045H9.32112C9.60204 16.1045 9.86706 16.1045 10.004 16.2646C10.0856 16.3602 10.1106 16.5021 10.0795 16.6982ZM14.5501 16.6803H13.8052C13.7417 16.6803 13.6869 16.7267 13.6771 16.7897L13.6441 16.9981L13.592 16.9226C13.4308 16.6885 13.0712 16.6103 12.7122 16.6103C11.889 16.6103 11.1859 17.2337 11.049 18.1084C10.9778 18.5446 11.079 18.9617 11.3264 19.2527C11.5535 19.5202 11.8783 19.6316 12.2647 19.6316C12.928 19.6316 13.2959 19.2051 13.2959 19.2051L13.2626 19.4122C13.2501 19.4913 13.3111 19.5623 13.3904 19.5623H14.0615C14.1682 19.5623 14.2582 19.4849 14.2751 19.3796L14.6777 16.83C14.6904 16.7515 14.6297 16.6803 14.5501 16.6803ZM13.5117 18.1302C13.4398 18.5557 13.102 18.8414 12.6713 18.8414C12.455 18.8414 12.282 18.7721 12.1711 18.6406C12.061 18.5101 12.0191 18.3242 12.0541 18.1172C12.1213 17.6952 12.4647 17.4003 12.889 17.4003C13.1004 17.4003 13.2724 17.4705 13.3857 17.6031C13.4992 17.7371 13.5442 17.9241 13.5117 18.1302ZM17.769 16.6802H18.5176C18.6224 16.6802 18.6836 16.7977 18.624 16.8837L16.1342 20.4776C16.0939 20.5359 16.0275 20.5705 15.9565 20.5705H15.2089C15.1036 20.5705 15.0421 20.452 15.1031 20.3657L15.8783 19.2714L15.0538 16.8516C15.0253 16.7675 15.0874 16.6802 15.1769 16.6802H15.9124C16.008 16.6802 16.0923 16.7429 16.1199 16.8344L16.5575 18.2958L17.59 16.775C17.6304 16.7156 17.6976 16.6802 17.769 16.6802Z"
+        fill="#253B80"
+      />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M25.8856 19.4129L26.5241 15.3509C26.5338 15.2878 26.5885 15.2414 26.652 15.2412H27.3707C27.4501 15.2412 27.511 15.3124 27.4985 15.3911L26.8689 19.38C26.8525 19.4852 26.762 19.5626 26.6553 19.5626H26.0134C25.9341 19.5626 25.8731 19.4916 25.8856 19.4129ZM20.9963 15.2414H19.4404C19.3341 15.2414 19.2436 15.3188 19.227 15.4238L18.5978 19.4129C18.5853 19.4916 18.6463 19.5626 18.7257 19.5626H19.524C19.5983 19.5626 19.6617 19.5085 19.6733 19.4348L19.8518 18.3041C19.8683 18.1987 19.959 18.1214 20.0652 18.1214H20.5574C21.5824 18.1214 22.1736 17.6255 22.3283 16.6429C22.398 16.213 22.331 15.8752 22.1297 15.6386C21.9088 15.3789 21.5169 15.2414 20.9963 15.2414ZM21.1757 16.6984C21.0909 17.2566 20.6644 17.2566 20.2517 17.2566H20.0172L20.1821 16.2141C20.1919 16.1511 20.246 16.1047 20.3099 16.1047H20.4175C20.6983 16.1047 20.9635 16.1047 21.1004 16.2648C21.182 16.3604 21.2069 16.5023 21.1757 16.6984ZM25.646 16.6805H24.9015C24.8375 16.6805 24.7832 16.7269 24.7736 16.79L24.7407 16.9983L24.6884 16.9228C24.5271 16.6887 24.1677 16.6105 23.8088 16.6105C22.9855 16.6105 22.2827 17.234 22.1457 18.1086C22.0748 18.5448 22.1755 18.962 22.423 19.2529C22.6505 19.5204 22.9749 19.6319 23.3613 19.6319C24.0246 19.6319 24.3924 19.2053 24.3924 19.2053L24.3592 19.4124C24.3467 19.4915 24.4076 19.5625 24.4875 19.5625H25.1583C25.2645 19.5625 25.355 19.4851 25.3716 19.3799L25.7745 16.8302C25.7868 16.7517 25.7258 16.6805 25.646 16.6805ZM24.6076 18.1304C24.5362 18.556 24.1979 18.8416 23.7671 18.8416C23.5512 18.8416 23.3779 18.7723 23.2669 18.6408C23.1568 18.5103 23.1154 18.3244 23.15 18.1174C23.2175 17.6954 23.5606 17.4005 23.9848 17.4005C24.1964 17.4005 24.3683 17.4707 24.4816 17.6033C24.5955 17.7373 24.6406 17.9243 24.6076 18.1304Z"
+        fill="#179BD7"
+      />
+      <path
+        d="M15.6571 13.8138L15.8484 12.5991L15.4224 12.5892H13.3882L14.8019 3.62534C14.8062 3.59828 14.8205 3.57305 14.8413 3.55513C14.8622 3.53722 14.8889 3.52734 14.9167 3.52734H18.3467C19.4855 3.52734 20.2713 3.7643 20.6816 4.232C20.8739 4.45141 20.9964 4.68068 21.0557 4.933C21.1178 5.19776 21.1189 5.51406 21.0582 5.89985L21.0538 5.92801V6.1752L21.2462 6.28418C21.4082 6.37011 21.5368 6.46848 21.6356 6.58111C21.8002 6.7687 21.9066 7.00712 21.9516 7.28978C21.998 7.5805 21.9826 7.92643 21.9066 8.31807C21.8188 8.76859 21.6769 9.16096 21.4854 9.48202C21.309 9.77785 21.0845 10.0232 20.818 10.2134C20.5635 10.394 20.261 10.5312 19.9191 10.619C19.5878 10.7052 19.2101 10.7488 18.7958 10.7488H18.5288C18.338 10.7488 18.1525 10.8175 18.007 10.9408C17.8611 11.0665 17.7645 11.2383 17.7349 11.4263L17.7148 11.5357L17.3769 13.6767L17.3616 13.7553C17.3576 13.7802 17.3506 13.7927 17.3404 13.8011C17.3312 13.8087 17.3181 13.8138 17.3052 13.8138H15.6571Z"
+        fill="#253B80"
+      />
+      <path
+        d="M21.4284 5.95703C21.4182 6.02248 21.4065 6.0894 21.3933 6.15815C20.9409 8.48056 19.3934 9.28286 17.4169 9.28286H16.4106C16.1689 9.28286 15.9652 9.45838 15.9275 9.6968L15.4123 12.9645L15.2664 13.8907C15.2418 14.0473 15.3626 14.1884 15.5205 14.1884H17.3054C17.5168 14.1884 17.6963 14.0348 17.7295 13.8264L17.7471 13.7357L18.0832 11.603L18.1048 11.4861C18.1376 11.2769 18.3175 11.1234 18.5289 11.1234H18.7958C20.5251 11.1234 21.8789 10.4213 22.2746 8.38951C22.4398 7.54077 22.3543 6.8321 21.9169 6.33367C21.7846 6.18338 21.6204 6.05868 21.4284 5.95703Z"
+        fill="#179BD7"
+      />
+      <path
+        d="M20.9549 5.7688C20.8859 5.74869 20.8145 5.7304 20.7414 5.71395C20.6679 5.69787 20.5925 5.6836 20.515 5.67117C20.2437 5.62729 19.9464 5.60645 19.6279 5.60645H16.9395C16.8732 5.60645 16.8104 5.62143 16.7541 5.6485C16.6301 5.7081 16.538 5.82549 16.5157 5.96919L15.9437 9.59159L15.9272 9.69727C15.9649 9.45884 16.1686 9.28332 16.4103 9.28332H17.4166C19.3931 9.28332 20.9407 8.48066 21.393 6.15862C21.4065 6.08987 21.4179 6.02295 21.4281 5.9575C21.3137 5.89679 21.1897 5.84487 21.0562 5.80061C21.0234 5.78965 20.9894 5.77905 20.9549 5.7688Z"
+        fill="#222D65"
+      />
+      <path
+        d="M16.5156 5.96866C16.5379 5.82495 16.6301 5.70757 16.754 5.64833C16.8107 5.62126 16.8733 5.60627 16.9394 5.60627H19.6279C19.9464 5.60627 20.2437 5.62712 20.5149 5.671C20.5925 5.68343 20.6679 5.6977 20.7414 5.71378C20.8145 5.73024 20.8858 5.74853 20.9549 5.76863C20.9893 5.77887 21.0233 5.78947 21.0565 5.80008C21.19 5.84433 21.314 5.89662 21.4284 5.95696C21.5631 5.09871 21.4274 4.51437 20.9634 3.98523C20.4518 3.40271 19.5284 3.15332 18.3469 3.15332H14.9168C14.6755 3.15332 14.4697 3.32884 14.4323 3.56763L13.0037 12.6236C12.9755 12.8028 13.1137 12.9645 13.2943 12.9645H15.412L15.9437 9.59105L16.5156 5.96866Z"
+        fill="#253B80"
+      />
+    </svg>
+  )
+}
+
+export const Visa = function Visa() {
+  return (
+    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="white" stroke="#D9D9D9" />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M10.6253 16.2582H8.50494L6.91495 10.1924C6.83949 9.91334 6.67925 9.66667 6.44354 9.5504C5.85531 9.25823 5.20712 9.0257 4.5 8.90843V8.6749H7.91567C8.38708 8.6749 8.74064 9.0257 8.79957 9.43313L9.62454 13.8086L11.7438 8.6749H13.8052L10.6253 16.2582ZM14.9838 16.2582H12.9813L14.6302 8.6749H16.6327L14.9838 16.2582ZM19.2234 10.7757C19.2823 10.3673 19.6359 10.1337 20.0483 10.1337C20.6965 10.0751 21.4026 10.1924 21.9919 10.4835L22.3454 8.85081C21.7562 8.61727 21.108 8.5 20.5198 8.5C18.5762 8.5 17.162 9.55041 17.162 11.0082C17.162 12.1173 18.1637 12.6996 18.8708 13.0504C19.6359 13.4002 19.9305 13.6338 19.8716 13.9835C19.8716 14.5082 19.2823 14.7418 18.6941 14.7418C17.9869 14.7418 17.2798 14.5669 16.6327 14.2747L16.2791 15.9085C16.9862 16.1996 17.7512 16.3169 18.4584 16.3169C20.6376 16.3745 21.9919 15.3251 21.9919 13.75C21.9919 11.7665 19.2234 11.6502 19.2234 10.7757ZM29 16.2582L27.41 8.6749H25.7022C25.3486 8.6749 24.9951 8.90843 24.8772 9.25823L21.9329 16.2582H23.9943L24.4058 15.1502H26.9386L27.1743 16.2582H29ZM25.9968 10.7171L26.585 13.5751H24.9361L25.9968 10.7171Z"
+        fill="#172B85"
+      />
+    </svg>
+  )
+}
