--- apps/web/vibes/soul/components/footer/payment-icons/bitcoin.tsx
@@ -1,15 +1,0 @@
-export const Bitcoin = function Bitcoin() {
-  return (
-    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="white" stroke="#D9D9D9" />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M25.6269 12.0646C25.6269 16.5185 22.0165 20.1292 17.5627 20.1292C13.109 20.1292 9.49854 16.5185 9.49854 12.0646C9.49854 7.61064 13.109 4 17.5627 4C22.0165 4 25.6269 7.61064 25.6269 12.0646ZM19.5001 9.06828C20.6219 9.45351 21.4424 10.0303 21.2815 11.1038C21.1645 11.8899 20.7272 12.2701 20.146 12.4032C20.9433 12.8167 21.215 13.6014 20.9626 14.5502C20.4832 15.9157 19.3437 16.0306 17.8281 15.7452L17.4601 17.213L16.5717 16.9922L16.9346 15.5441C16.7043 15.4871 16.4689 15.4261 16.2265 15.361L15.862 16.8162L14.9746 16.5954L15.3421 15.1245L13.5527 14.6748L13.9941 13.6613C13.9941 13.6613 14.6491 13.8342 14.64 13.8216C14.8916 13.8836 15.0035 13.7203 15.0475 13.6119L16.0447 9.62797C16.0559 9.43989 15.9906 9.20291 15.6307 9.11316C15.6443 9.10358 14.9852 8.95332 14.9852 8.95332L15.2216 8.00789L17.0136 8.4501L17.3781 6.99591L18.266 7.21676L17.9091 8.64221C18.148 8.69616 18.388 8.75112 18.6213 8.80911L18.9757 7.39274L19.8641 7.61359L19.5001 9.06828ZM17.3744 11.4941C17.9796 11.6545 19.2964 12.0037 19.5259 11.0882C19.7601 10.1516 18.4807 9.86878 17.8543 9.73028L17.8543 9.73027C17.7833 9.71457 17.7207 9.70074 17.6697 9.68801L17.2272 11.4558C17.2694 11.4662 17.3189 11.4794 17.3744 11.4941ZM16.6887 14.3404C17.4136 14.5313 18.9988 14.9489 19.251 13.9402C19.5092 12.9088 17.9722 12.5652 17.2223 12.3976C17.1385 12.3789 17.0645 12.3624 17.004 12.3473L16.516 14.2956C16.5656 14.3079 16.6236 14.3232 16.6886 14.3403L16.6887 14.3404Z"
-        fill="#F7931A"
-      />
-    </svg>
-  )
-}
-
-export default Bitcoin
