--- apps/web/vibes/soul/components/footer/payment-icons/google-pay.tsx
@@ -1,31 +1,0 @@
-export const GooglePay = function GooglePay() {
-  return (
-    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="white" stroke="#D9D9D9" />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M16.5301 15.7585V12.7375H18.0893C18.7282 12.7375 19.2675 12.5234 19.7071 12.1012L19.8126 11.9942C20.6156 11.12 20.5629 9.75817 19.7071 8.94941C19.2792 8.52125 18.6931 8.28932 18.0893 8.30122H15.5864V15.7585H16.5301ZM16.5303 11.8217V9.21703H18.1131C18.4531 9.21703 18.7756 9.34786 19.0159 9.58573C19.526 10.0853 19.5377 10.9178 19.0453 11.4352C18.8049 11.6909 18.4649 11.8336 18.1131 11.8217H16.5303ZM24.2147 11.0546C23.8102 10.6799 23.2592 10.4896 22.5617 10.4896C21.6649 10.4896 20.9908 10.8227 20.5453 11.4827L21.3777 12.012C21.6825 11.5601 22.0987 11.3341 22.6262 11.3341C22.9603 11.3341 23.2827 11.459 23.5347 11.6849C23.7809 11.899 23.9216 12.2083 23.9216 12.5353V12.7554C23.5582 12.5532 23.101 12.4461 22.5383 12.4461C21.8818 12.4461 21.3542 12.6007 20.9615 12.9159C20.5688 13.2311 20.3695 13.6474 20.3695 14.1766C20.3578 14.6583 20.5629 15.1162 20.9263 15.4254C21.2956 15.7585 21.7645 15.925 22.3155 15.925C22.9662 15.925 23.482 15.6336 23.8747 15.0508H23.9157V15.7585H24.8184V12.6126C24.8184 11.9525 24.6191 11.4292 24.2147 11.0546ZM21.6533 14.8189C21.4599 14.6761 21.3426 14.4442 21.3426 14.1944C21.3426 13.9149 21.4716 13.683 21.7237 13.4987C21.9816 13.3143 22.3041 13.2192 22.6851 13.2192C23.2128 13.2132 23.6231 13.3322 23.9163 13.57C23.9163 13.9744 23.758 14.3253 23.4473 14.6226C23.1659 14.9081 22.7848 15.0686 22.3862 15.0686C22.1223 15.0746 21.8644 14.9854 21.6533 14.8189ZM26.8465 18.0004L30 10.6562H28.9743L27.5147 14.3194H27.4972L26.0025 10.6562H24.9767L27.0458 15.4314L25.8735 18.0004H26.8465Z"
-        fill="#3C4043"
-      />
-      <path
-        d="M13.272 12.0831C13.272 11.7917 13.2486 11.5003 13.2017 11.2148H9.22168V12.8621H11.5018C11.408 13.3914 11.1032 13.8671 10.6578 14.1644V15.2349H12.0176C12.8148 14.4915 13.272 13.3914 13.272 12.0831Z"
-        fill="#4285F4"
-      />
-      <path
-        d="M9.22211 16.2696C10.3592 16.2696 11.3205 15.889 12.0181 15.2349L10.6582 14.1645C10.2772 14.4261 9.79068 14.5748 9.22211 14.5748C8.12013 14.5748 7.18815 13.8195 6.85404 12.8086H5.45312V13.9147C6.16824 15.3598 7.62776 16.2696 9.22211 16.2696Z"
-        fill="#34A853"
-      />
-      <path
-        d="M6.85401 12.8088C6.67814 12.2796 6.67814 11.7027 6.85401 11.1675V10.0674H5.45288C4.84904 11.2746 4.84904 12.7018 5.45288 13.909L6.85401 12.8088Z"
-        fill="#FBBC04"
-      />
-      <path
-        d="M9.22211 9.40114C9.82585 9.38925 10.4061 9.62117 10.8399 10.0434L12.0474 8.81836C11.2795 8.09285 10.2713 7.69442 9.22211 7.70631C7.62776 7.70631 6.16824 8.62211 5.45312 10.0672L6.85404 11.1733C7.18815 10.1564 8.12013 9.40114 9.22211 9.40114Z"
-        fill="#EA4335"
-      />
-    </svg>
-  )
-}
-
-export default GooglePay
