--- apps/web/vibes/soul/components/footer/payment-icons/paypal.tsx
@@ -1,37 +1,0 @@
-export const Paypal = function Paypal() {
-  return (
-    <svg width="35" height="24" viewBox="0 0 35 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <rect x="0.5" y="0.5" width="34" height="23" rx="3.5" fill="white" stroke="#D9D9D9" />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M9.90001 15.2412H8.34439C8.23794 15.2412 8.1474 15.3186 8.1308 15.4236L7.50163 19.4127C7.48913 19.4914 7.55009 19.5624 7.62993 19.5624H8.37259C8.47905 19.5624 8.56958 19.485 8.58618 19.3797L8.75588 18.3038C8.77225 18.1985 8.863 18.1212 8.96924 18.1212H9.46169C10.4864 18.1212 11.0778 17.6253 11.2323 16.6427C11.3019 16.2128 11.2353 15.875 11.0339 15.6384C10.8129 15.3787 10.4206 15.2412 9.90001 15.2412ZM10.0795 16.6982C9.99443 17.2564 9.56792 17.2564 9.15553 17.2564H8.92078L9.08547 16.2138C9.09525 16.1509 9.14984 16.1045 9.21353 16.1045H9.32112C9.60204 16.1045 9.86706 16.1045 10.004 16.2646C10.0856 16.3602 10.1106 16.5021 10.0795 16.6982ZM14.5501 16.6803H13.8052C13.7417 16.6803 13.6869 16.7267 13.6771 16.7897L13.6441 16.9981L13.592 16.9226C13.4308 16.6885 13.0712 16.6103 12.7122 16.6103C11.889 16.6103 11.1859 17.2337 11.049 18.1084C10.9778 18.5446 11.079 18.9617 11.3264 19.2527C11.5535 19.5202 11.8783 19.6316 12.2647 19.6316C12.928 19.6316 13.2959 19.2051 13.2959 19.2051L13.2626 19.4122C13.2501 19.4913 13.3111 19.5623 13.3904 19.5623H14.0615C14.1682 19.5623 14.2582 19.4849 14.2751 19.3796L14.6777 16.83C14.6904 16.7515 14.6297 16.6803 14.5501 16.6803ZM13.5117 18.1302C13.4398 18.5557 13.102 18.8414 12.6713 18.8414C12.455 18.8414 12.282 18.7721 12.1711 18.6406C12.061 18.5101 12.0191 18.3242 12.0541 18.1172C12.1213 17.6952 12.4647 17.4003 12.889 17.4003C13.1004 17.4003 13.2724 17.4705 13.3857 17.6031C13.4992 17.7371 13.5442 17.9241 13.5117 18.1302ZM17.769 16.6802H18.5176C18.6224 16.6802 18.6836 16.7977 18.624 16.8837L16.1342 20.4776C16.0939 20.5359 16.0275 20.5705 15.9565 20.5705H15.2089C15.1036 20.5705 15.0421 20.452 15.1031 20.3657L15.8783 19.2714L15.0538 16.8516C15.0253 16.7675 15.0874 16.6802 15.1769 16.6802H15.9124C16.008 16.6802 16.0923 16.7429 16.1199 16.8344L16.5575 18.2958L17.59 16.775C17.6304 16.7156 17.6976 16.6802 17.769 16.6802Z"
-        fill="#253B80"
-      />
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M25.8856 19.4129L26.5241 15.3509C26.5338 15.2878 26.5885 15.2414 26.652 15.2412H27.3707C27.4501 15.2412 27.511 15.3124 27.4985 15.3911L26.8689 19.38C26.8525 19.4852 26.762 19.5626 26.6553 19.5626H26.0134C25.9341 19.5626 25.8731 19.4916 25.8856 19.4129ZM20.9963 15.2414H19.4404C19.3341 15.2414 19.2436 15.3188 19.227 15.4238L18.5978 19.4129C18.5853 19.4916 18.6463 19.5626 18.7257 19.5626H19.524C19.5983 19.5626 19.6617 19.5085 19.6733 19.4348L19.8518 18.3041C19.8683 18.1987 19.959 18.1214 20.0652 18.1214H20.5574C21.5824 18.1214 22.1736 17.6255 22.3283 16.6429C22.398 16.213 22.331 15.8752 22.1297 15.6386C21.9088 15.3789 21.5169 15.2414 20.9963 15.2414ZM21.1757 16.6984C21.0909 17.2566 20.6644 17.2566 20.2517 17.2566H20.0172L20.1821 16.2141C20.1919 16.1511 20.246 16.1047 20.3099 16.1047H20.4175C20.6983 16.1047 20.9635 16.1047 21.1004 16.2648C21.182 16.3604 21.2069 16.5023 21.1757 16.6984ZM25.646 16.6805H24.9015C24.8375 16.6805 24.7832 16.7269 24.7736 16.79L24.7407 16.9983L24.6884 16.9228C24.5271 16.6887 24.1677 16.6105 23.8088 16.6105C22.9855 16.6105 22.2827 17.234 22.1457 18.1086C22.0748 18.5448 22.1755 18.962 22.423 19.2529C22.6505 19.5204 22.9749 19.6319 23.3613 19.6319C24.0246 19.6319 24.3924 19.2053 24.3924 19.2053L24.3592 19.4124C24.3467 19.4915 24.4076 19.5625 24.4875 19.5625H25.1583C25.2645 19.5625 25.355 19.4851 25.3716 19.3799L25.7745 16.8302C25.7868 16.7517 25.7258 16.6805 25.646 16.6805ZM24.6076 18.1304C24.5362 18.556 24.1979 18.8416 23.7671 18.8416C23.5512 18.8416 23.3779 18.7723 23.2669 18.6408C23.1568 18.5103 23.1154 18.3244 23.15 18.1174C23.2175 17.6954 23.5606 17.4005 23.9848 17.4005C24.1964 17.4005 24.3683 17.4707 24.4816 17.6033C24.5955 17.7373 24.6406 17.9243 24.6076 18.1304Z"
-        fill="#179BD7"
-      />
-      <path
-        d="M15.6571 13.8138L15.8484 12.5991L15.4224 12.5892H13.3882L14.8019 3.62534C14.8062 3.59828 14.8205 3.57305 14.8413 3.55513C14.8622 3.53722 14.8889 3.52734 14.9167 3.52734H18.3467C19.4855 3.52734 20.2713 3.7643 20.6816 4.232C20.8739 4.45141 20.9964 4.68068 21.0557 4.933C21.1178 5.19776 21.1189 5.51406 21.0582 5.89985L21.0538 5.92801V6.1752L21.2462 6.28418C21.4082 6.37011 21.5368 6.46848 21.6356 6.58111C21.8002 6.7687 21.9066 7.00712 21.9516 7.28978C21.998 7.5805 21.9826 7.92643 21.9066 8.31807C21.8188 8.76859 21.6769 9.16096 21.4854 9.48202C21.309 9.77785 21.0845 10.0232 20.818 10.2134C20.5635 10.394 20.261 10.5312 19.9191 10.619C19.5878 10.7052 19.2101 10.7488 18.7958 10.7488H18.5288C18.338 10.7488 18.1525 10.8175 18.007 10.9408C17.8611 11.0665 17.7645 11.2383 17.7349 11.4263L17.7148 11.5357L17.3769 13.6767L17.3616 13.7553C17.3576 13.7802 17.3506 13.7927 17.3404 13.8011C17.3312 13.8087 17.3181 13.8138 17.3052 13.8138H15.6571Z"
-        fill="#253B80"
-      />
-      <path
-        d="M21.4284 5.95703C21.4182 6.02248 21.4065 6.0894 21.3933 6.15815C20.9409 8.48056 19.3934 9.28286 17.4169 9.28286H16.4106C16.1689 9.28286 15.9652 9.45838 15.9275 9.6968L15.4123 12.9645L15.2664 13.8907C15.2418 14.0473 15.3626 14.1884 15.5205 14.1884H17.3054C17.5168 14.1884 17.6963 14.0348 17.7295 13.8264L17.7471 13.7357L18.0832 11.603L18.1048 11.4861C18.1376 11.2769 18.3175 11.1234 18.5289 11.1234H18.7958C20.5251 11.1234 21.8789 10.4213 22.2746 8.38951C22.4398 7.54077 22.3543 6.8321 21.9169 6.33367C21.7846 6.18338 21.6204 6.05868 21.4284 5.95703Z"
-        fill="#179BD7"
-      />
-      <path
-        d="M20.9549 5.7688C20.8859 5.74869 20.8145 5.7304 20.7414 5.71395C20.6679 5.69787 20.5925 5.6836 20.515 5.67117C20.2437 5.62729 19.9464 5.60645 19.6279 5.60645H16.9395C16.8732 5.60645 16.8104 5.62143 16.7541 5.6485C16.6301 5.7081 16.538 5.82549 16.5157 5.96919L15.9437 9.59159L15.9272 9.69727C15.9649 9.45884 16.1686 9.28332 16.4103 9.28332H17.4166C19.3931 9.28332 20.9407 8.48066 21.393 6.15862C21.4065 6.08987 21.4179 6.02295 21.4281 5.9575C21.3137 5.89679 21.1897 5.84487 21.0562 5.80061C21.0234 5.78965 20.9894 5.77905 20.9549 5.7688Z"
-        fill="#222D65"
-      />
-      <path
-        d="M16.5156 5.96866C16.5379 5.82495 16.6301 5.70757 16.754 5.64833C16.8107 5.62126 16.8733 5.60627 16.9394 5.60627H19.6279C19.9464 5.60627 20.2437 5.62712 20.5149 5.671C20.5925 5.68343 20.6679 5.6977 20.7414 5.71378C20.8145 5.73024 20.8858 5.74853 20.9549 5.76863C20.9893 5.77887 21.0233 5.78947 21.0565 5.80008C21.19 5.84433 21.314 5.89662 21.4284 5.95696C21.5631 5.09871 21.4274 4.51437 20.9634 3.98523C20.4518 3.40271 19.5284 3.15332 18.3469 3.15332H14.9168C14.6755 3.15332 14.4697 3.32884 14.4323 3.56763L13.0037 12.6236C12.9755 12.8028 13.1137 12.9645 13.2943 12.9645H15.412L15.9437 9.59105L16.5156 5.96866Z"
-        fill="#253B80"
-      />
-    </svg>
-  )
-}
-
-export default Paypal
