+++ apps/web/vibes/soul/components/page-catalyst-home/index.tsx
@@ -2,16 +2,16 @@ import FeaturedImage from '@/vibes/soul/components/featured-image'
 import FeaturedProductsCarousel from '@/vibes/soul/components/featured-products-carousel'
 import FeaturedProductsList from '@/vibes/soul/components/featured-products-list'
 import Footer from '@/vibes/soul/components/footer'
-import Amex from '@/vibes/soul/components/footer/payment-icons/amex'
-import ApplePay from '@/vibes/soul/components/footer/payment-icons/apple-pay'
-import Bitcoin from '@/vibes/soul/components/footer/payment-icons/bitcoin'
-import GooglePay from '@/vibes/soul/components/footer/payment-icons/google-pay'
-import Mastercard from '@/vibes/soul/components/footer/payment-icons/mastercard'
-import Paypal from '@/vibes/soul/components/footer/payment-icons/paypal'
-import Visa from '@/vibes/soul/components/footer/payment-icons/visa'
-import Facebook from '@/vibes/soul/components/footer/social-icons/facebook'
-import LinkedIn from '@/vibes/soul/components/footer/social-icons/linked-in'
-import X from '@/vibes/soul/components/footer/social-icons/x'
+import {
+  Amex,
+  ApplePay,
+  Bitcoin,
+  GooglePay,
+  Mastercard,
+  Paypal,
+  Visa,
+} from '@/vibes/soul/components/footer/payment-icons'
+import { Facebook, Instagram, X, Youtube } from '@/vibes/soul/components/footer/social-icons'
 import Header from '@/vibes/soul/components/header'
 import Slideshow from '@/vibes/soul/components/slideshow'
 import SubscribeBasic from '@/vibes/soul/components/subscribe-basic'
@@ -25,11 +25,15 @@ const socialMediaLinks = [
   },
   {
     href: '#',
-    icon: <LinkedIn />,
+    icon: <X />,
   },
   {
     href: '#',
-    icon: <X />,
+    icon: <Instagram />,
+  },
+  {
+    href: '#',
+    icon: <Youtube />,
   },
 ]
